requirements.txt
django
python-telegram-bot
openai
celery
redis
python-decouple
django-environ
psycopg2-binary
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
APScheduler
pytz
python-dotenv
gunicorn
django-celery-beat
whitenoise